+++ orange/src/components/index.tsx
@@ -8,7 +8,6 @@ import NecessityTemplate from './Necessity/NecessityTemplate/NecessityTemplate';
 import SignUpModal from './SignUpModal/SignUpModal';
 import StudyInfo from './StudyInfo/StudyInfo';
 import LogList from './LogList/LogList';
-import HouseList from './HouseList/HouseList';
 
 export {
   Header,
@@ -21,5 +20,4 @@ export {
   SignUpModal,
   StudyInfo,
   LogList,
-  HouseList,
 };
